Divided feature into components
