+++ src/features/admin/refund_requests/EvaluateRefunds.js
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from "react";
 import RefundService from "services/RefundService";
-import { message, Table, Button } from "antd";
+import { message } from "antd";
+import RefundTable from "./components/RefundTable";
 
 const EvaluateRefunds = () => {
   const [refundRequests, setRefundRequests] = useState([]);
@@ -34,7 +35,21 @@ const EvaluateRefunds = () => {
       message.error("Failed to update refund status");
     }
   };
-  
+
+  const isOlderThan30Days = (orderDate) => {
+    const orderDateObj = new Date(orderDate);
+    const currentDate = new Date();
+    const diffTime = Math.abs(currentDate.getTime() - orderDateObj.getTime());
+    const diffDays = Math.ceil(diffTime / (1000 * 60 * 60 * 24));
+    return diffDays > 30;
+  };
+
+  refundRequests.forEach((request) => {
+    if (isOlderThan30Days(request.orderDate) && request.status === 'PENDING') {
+      handleUpdateStatus(request, 'REJECTED');
+    }
+  });
+
   if (loading) {
     return <div>Loading...</div>;
   }
@@ -43,58 +58,11 @@ const EvaluateRefunds = () => {
     return <div>{error}</div>;
   }
 
-const isOlderThan30Days = (orderDate) => {
-    const orderDateObj = new Date(orderDate);
-    const currentDate = new Date();
-    const diffTime = Math.abs(currentDate.getTime() - orderDateObj.getTime());
-    const diffDays = Math.ceil(diffTime / (1000 * 60 * 60 * 24));
-    return diffDays > 30;
-};
-
-refundRequests.forEach((request) => {
-    if (isOlderThan30Days(request.orderDate) && request.status === 'PENDING') {
-        handleUpdateStatus(request, 'REJECTED');
-    }
-});
-
-return (
+  return (
     <div>
-        <Table dataSource={refundRequests} rowKey="refundRequestId">
-            <Table.Column title="Order ID" dataIndex="orderId" key="orderId" />
-            <Table.Column title="Product ID" dataIndex="productId" key="productId" />
-            <Table.Column title="Product Name" dataIndex="productName" key="productName" />
-            <Table.Column title="User Email" dataIndex="userEmail" key="userEmail" />
-            <Table.Column title="Order Date" dataIndex="orderDate" key="orderDate" />
-            <Table.Column title="Reason" dataIndex="reason" key="reason" />
-            <Table.Column title="Status" dataIndex="status" key="status" />
-            <Table.Column
-                title="Actions"
-                key="actions"
-                render={(text, record) => (
-                    <span>
-                        <Button
-                            style={{ marginRight: 8 }}
-                            color="primary"
-                            variant="solid"
-                            onClick={() => handleUpdateStatus(record, 'APPROVED')}
-                            disabled={record.status !== 'PENDING'}
-                        >
-                            Approve
-                        </Button>
-                        <Button
-                            color="danger"
-                            variant="solid"
-                            onClick={() => handleUpdateStatus(record, 'REJECTED')}
-                            disabled={record.status !== 'PENDING'}
-                        >
-                            Reject
-                        </Button>
-                    </span>
-                )}
-            />
-        </Table>
+      <RefundTable refundRequests={refundRequests} handleUpdateStatus={handleUpdateStatus} />
     </div>
-);
+  );
 };
 
 export default EvaluateRefunds;
